pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "uk-biz-intel-mcp"
version = "1.0.0"
description = "UK Business Intelligence MCP server — unified due diligence across five public registries"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["mcp", "uk", "companies-house", "due-diligence", "fastmcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
uk-biz-intel-mcp = "server:main"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true
